rename `h5p` to `H5P`
