+++ src/components/AppearanceDialog/AppearanceDialog.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
-import { t } from "../../h5p/H5P.util";
+import { t } from "../../H5P/H5P.util";
 import { ColorTheme } from "../../types/ColorTheme";
-import { H5PFieldImage } from "../../types/h5p/H5PField";
-import { H5PForm } from "../../types/h5p/H5PForm";
-import { Params } from "../../types/h5p/Params";
+import { H5PFieldImage } from "../../types/H5P/H5PField";
+import { H5PForm } from "../../types/H5P/H5PForm";
+import { Params } from "../../types/H5P/Params";
 import { defaultTheme } from "../../utils/theme.utils";
 import { Dialog } from "../Dialog/Dialog";
 import { SemanticsForm } from "../SemanticsForm/SemanticsForm";
