+++ src/components/MapEditorView/MapEditorView.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
-import { t } from "../../h5p/H5P.util";
+import { t } from "../../H5P/H5P.util";
 import { ArrowItemType } from "../../types/ArrowItemType";
-import { H5PFieldGroup } from "../../types/h5p/H5PField";
-import { H5PForm } from "../../types/h5p/H5PForm";
-import { Params } from "../../types/h5p/Params";
+import { H5PFieldGroup } from "../../types/H5P/H5PField";
+import { H5PForm } from "../../types/H5P/H5PForm";
+import { Params } from "../../types/H5P/Params";
 import { TopicMapItemType } from "../../types/TopicMapItemType";
 import { getBackgroundImageField } from "../../utils/H5P/form.utils";
 import { Dialog } from "../Dialog/Dialog";
