+++ src/index.tsx
@@ -1,5 +1,5 @@
-import { H5PEditor } from "./h5p/H5P.util";
-import { H5PWrapper } from "./h5p/H5PWrapper";
+import { H5PEditor } from "./H5P/H5P.util";
+import { H5PWrapper } from "./H5P/H5PWrapper";
 import "./styles.scss";
 
 H5PEditor.widgets.topicMap = H5PWrapper;
