+++ src/types/Appearance.ts
@@ -1,5 +1,5 @@
 import { ColorTheme } from "./ColorTheme";
-import { Image } from "./h5p/Image";
+import { Image } from "./H5P/Image";
 
 export type Appearance = {
   backgroundImage?: Image;
