+++ src/types/CommonItemType.ts
@@ -1,5 +1,5 @@
 import { DialogContent } from "./DialogContent";
-import { Image } from "./h5p/Image";
+import { Image } from "./H5P/Image";
 
 export type CommonItemType = {
   id: string;
