+++ src/types/H5P/Copyright.ts
