+++ src/types/H5P/H5PEnterMode.ts
