+++ src/types/H5P/Image.ts
